refactor(settings): split SettingsPage into per-panel components
SettingsPage.tsx grew to 2210 lines with 14 panels and helpers all colocated.
Extract each panel into src/components/settings/ for testability and clearer
ownership. SettingsPage now acts as a router of hash-based panel selection.

Bonus fix: missing `supabase` import in SecuritePanel was silently broken
(used at runtime via implicit global). Now properly imported from
@/lib/supabase/client.

- src/pages/SettingsPage.tsx: 2210 → 87 lines (-96%)
- New: src/components/settings/{ProfilPanel, SecuritePanel, AbonnementPanel,
  NotificationsPanel, InterventionsPanel, ConventionPanel, PchPanel,
  ApparencePanel, AccessibilitePanel, DonneesPanel}.tsx
- New: SettingsNavigation.tsx (sidebar) + navigationConfig.tsx (sections + types)
- New: SettingsShared.tsx (PanelHeader, ToggleRow)
- Cards stay inline in their parent panel (TwoFactorCard + DangerZone in
  SecuritePanel, UsageBar in AbonnementPanel, HealthConsentCard +
  PrivacySettingsCard in DonneesPanel)
- No behavior change. All 22 existing SettingsPage tests still pass.
