+++ src/components/settings/index.ts
@@ -1,0 +1,14 @@
+export { SettingsNavigation } from './SettingsNavigation'
+export { panelFromHash, NAV_SECTIONS } from './navigationConfig'
+export type { PanelId, NavItem, NavSection } from './navigationConfig'
+export { PanelHeader, ToggleRow } from './SettingsShared'
+export { ProfilPanel } from './ProfilPanel'
+export { SecuritePanel } from './SecuritePanel'
+export { AbonnementPanel } from './AbonnementPanel'
+export { NotificationsPanel } from './NotificationsPanel'
+export { InterventionsPanel } from './InterventionsPanel'
+export { ConventionPanel } from './ConventionPanel'
+export { PchPanel } from './PchPanel'
+export { ApparencePanel } from './ApparencePanel'
+export { AccessibilitePanel } from './AccessibilitePanel'
+export { DonneesPanel } from './DonneesPanel'
